FIX1/3: processRectanglePrompt() got global but 2D offset:
At the end of the prompting drag event, a 2D global px coordinate of
the created prompt is established (into a 'box' Interval), this box
should be fitting to the annotation site view (image crop; 'viewImg'),
and indeed it is as long as one understands the the 'viewImg' has non-zero
min() bound. SAM network assume a plain image (with min() bound at zeros),
so 'box' coords must be relativized to such an image, which is the reason
why two new params were added to the processRectanglePrompt().