CHG1/2: annotation site = SpatioTemporalView and that's it, consequently:
The new concept creates annotation sites automatically, on demand just
by drawing the first (rectangular) prompt. The system monitors if a view
has changed and if so, it flags that "nothing is valid anymore" and "needs
to be prepared again for the current view". This is achieved with the new
'isNextPromptOnNewAnnotationSite' flag, and "valid anymore" basically
refers to the content of the 'annotationSiteViewImg', which acts as a cache
for submitting the new prompts as long as still the same annotation site
is used.