BUGFIX on BdvPrompts with SourceAndConverter as operateOnThisSource:
The thing was that the viewer.state() transform is between
the screen coordinates and the implicit BDV's internal 3D space,
which is not necessarily identical to the coordinate space
of the selected image (image source). So, the "source transform"
needs to be kept next to the this.image and is used when
collectViewPixelData() (to, again, reach the correct pixel coordinate
from the queried real implicit internal BDV's 3D coordinates).

Besides, switchToThisSource() and both c'tor started to rely
on this and switchToThisImage() functions.