ADD: util/Prompts.findSeedsAndReturnAsBoxes() to find seeds for SAM, and apply SAM:
Considering how user carefully adjusted contrast in BDV, the contrast-adjusted
image is thresholded, decently closed (morphology), CCA, and boundig boxes are
established -- the boxes are with some simple statistics about pixel values
that are(!) taken from the original image (not from the contrast-adjusted one).

The boxes are then filtered for size and(!) for having included a very bright
pixel within the box (this turned out to be the crucial feature/factor).

And only the boxes that passed through are used as prompts to the SAM network,
which again operates on the original (not contrast-adjusted) image.

The new util class took over the heavy lifting of the "seeds" detection,
it does exactly what is described above.

And the BdvPrompts.thresholdAndProcessRectanglePrompt() is adapted to this API.