ADD,CHG2/2: util/Prompts.findSeedsAndReturnAsBoxes() that...
took over the heavy lifting of the "seeds" detection. It does exactly
the same algorithm that was introduced in the previous commit: (in fact
local) thresholding, small closing, CCA, filtering bounding boxes based
on their maximum intensities. It produces list of (hopefully) valid boxes
that are further processed (e.g. taken as individual prompts one-by-one
and subject to SAM network).

The BdvPrompts.thresholdAndProcessRectanglePrompt() adapted to this API.