(BUG)FIX: stop drawing prompt rectangle after (tracking)slicing by...
...removing the hack with samjOverlay.setStartOfLine() that
sets samjOverlay.isLineReadyForDrawing = false and also changes
TL corner of the prompt (which is unwanted in this case). Soo,
samjOverlay.requestNoDrawingOfPromptLine() was added

Also the request for final repainting (to e.g. not display the
prompt rectangle) was moved to BdvPrompts3D.run()'s end rather than
doing it (again!!) with every slice in DragBehaviourSkeletonFor3D.handleSlice()