FIX1/2: multikeys (shift+key) dragging session as critical section...
...which is making sure that atmost one session is working
with BdvPrompts. This needed to be implemented because
Scijava DragBehaviour allows to have several drag-sessions side
by side, not to mention that multikeys session typically run
even when no keys is pressed (drag session end is not detected
reliably -> see the next commit).

So, while we cannot prohibit to to have several drag-sessions
with Scijava, by introducing the critical section (in the form
of AtomicBoolean 'isAnybodyInDragSession' flag) we're making
sure that only one drag-session is actually doing something,
and the remaining sessions are void/doing nothing.