FIX2/2: own monitoring of a drag session end via KeyStrokesMonitor..
... because Scijava DragBehaviour has issues monitoring multi-key
(with shift-key pressed) sessions, we had to attach own monitor
of keys pressed and this one is queried during the drag session.

When no keys are detected, the session's end() method is called,
while the session for Scijava DragBehaviour keeps going on and
finishes typically with the end of another drag session. This would
trigger a call to end() again but the class is ready for that (uses
its own 'isMyDragSessionValid' flag).

The key monitor is registered with viewer panel in BdvPrompts.installBasicBehaviours(),
which is called from every BdvPrompts c'tor.